FARRAGO/FENNEL - Use unsigned 8-byte ints instead of 4-byte signed ints to        pass stats information for the hash join, hash agg, and sort exec        streams
[git-p4: depot-paths = "//open/lu/dev/": change = 7553]

+++ farrago/src/com/lucidera/opt/LhxAggRel.java
@@ -44,12 +44,12 @@ public class LhxAggRel
     /**
      * row count of the input
      */
-    int numInputRows;
+    long numInputRows;
 
     /**
      * cardinality of the group by key
      */
-    int cndGroupByKey;
+    long cndGroupByKey;
 
     //~ Constructors -----------------------------------------------------------
 
@@ -58,8 +58,8 @@ public class LhxAggRel
         RelNode child,
         int groupCount,
         Call [] aggCalls,
-        int numInputRows,
-        int cndGroupByKey)
+        long numInputRows,
+        long cndGroupByKey)
     {
         super(
             cluster,
